% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{get_garmin_data}
\alias{get_garmin_data}
\title{Log in to garmin connect}
\usage{
get_garmin_data(
  username,
  password,
  start_date,
  end_date,
  number_of_activities,
  verbose = TRUE,
  path = NULL
)
}
\arguments{
\item{username}{character, your Garmin Connect username (usuallly email address)}

\item{password}{character, your Garmin Connect password}

\item{start_date}{character, date from which on monitoring data should be collected}

\item{end_date}{character, date by which monitoring data should be collected}

\item{number_of_activities}{integer, number of activities to download}

\item{verbose}{boolean, display progress bar? (default = TRUE)}

\item{path}{character, where to create the databases to store the data (default = current working directory)}
}
\value{

}
\description{
Log in to garmin connect
}
